Renamed Adobe removals to be more general:
script removes target file on computer as given in Parameter 4 in JAMF, as root user. This concept is the same for the uninstaller, which can be given a parameter in JAMF to attempt an uninstall, but care must be taken to know if additional flags are needed for silent running.